pyproject.toml
[project]
name = "ai-manufacturing-workspace"
version = "0.1.0"
description = "Manufacturing document workspace backend scaffolding"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13",
  "cryptography>=42.0",
  "fastapi>=0.115",
  "httpx>=0.27",
  "pydantic>=2.7",
  "python-multipart>=0.0.9",
  "PyJWT[crypto]>=2.8",
  "PyYAML>=6.0",
  "SQLAlchemy>=2.0",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
processing = [
  "ezdxf>=1.3",
  "openpyxl>=3.1",
  "Pillow>=10.0",
  "PyMuPDF>=1.24",
  "python-docx>=1.1",
]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
